+++ src/shared/components/case-viewer/case-full-access-view/case-full-access-view.component.ts
@@ -10,7 +10,7 @@ import { Subscription } from 'rxjs/Subscription';
 
 import { DeleteOrCancelDialogComponent } from '../../dialogs';
 import { ShowCondition } from '../../../directives';
-import { Activity, CaseField, CaseTab, CaseView, CaseViewTrigger, DisplayMode, Draft, DRAFT_QUERY_PARAM, } from '../../../domain';
+import { Activity, CaseField, CaseTab, CaseView, CaseViewTrigger, DisplayMode, Draft, DRAFT_QUERY_PARAM } from '../../../domain';
 import {
   ActivityPollingService,
   AlertService,
@@ -19,6 +19,7 @@ import {
   NavigationNotifierService,
   NavigationOrigin,
   OrderService,
+  SessionStorageService
 } from '../../../services';
 import { CallbackErrorsContext } from '../../error';
 import { initDialog } from '../../helpers';
@@ -74,7 +75,8 @@ export class CaseFullAccessViewComponent implements OnInit, OnDestroy, AfterView
     private readonly draftService: DraftService,
     private readonly errorNotifierService: ErrorNotifierService,
     private convertHrefToRouterService: ConvertHrefToRouterService,
-    private readonly location: Location
+    private readonly location: Location,
+    private readonly sessionStorageService: SessionStorageService
   ) {
   }
 
@@ -93,6 +95,8 @@ export class CaseFullAccessViewComponent implements OnInit, OnDestroy, AfterView
       }
     });
     this.markdownUseHrefAsRouterLink = true;
+
+    this.sessionStorageService.removeItem('eventUrl');
 
     this.subscription = this.convertHrefToRouterService.getHrefMarkdownLinkContent().subscribe((hrefMarkdownLinkContent: string) => {
       // do not convert router with initial default value; convert to router only on updated link content