+++ src/shared/components/case-viewer/case-viewer.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs/Subscription';
 import { AbstractAppConfig } from '../../../app.config';
-import { CaseField, CaseTab, CaseView } from '../../domain';
+import { CaseTab, CaseView } from '../../domain';
 import { CaseNotifier } from '../case-editor';
 
 @Component({
@@ -12,7 +12,9 @@ import { CaseNotifier } from '../case-editor';
 export class CaseViewerComponent implements OnInit, OnDestroy {
 
   static readonly METADATA_FIELD_ACCESS_PROCEES_ID = '[ACCESS_PROCESS]';
+  static readonly METADATA_FIELD_ACCESS_GRANTED_ID = '[ACCESS_GRANTED]';
   static readonly NON_STANDARD_USER_ACCESS_TYPES = ['CHALLENGED', 'SPECIFIC'];
+  static readonly BASIC_USER_ACCESS_TYPES = 'BASIC';
 
   @Input() public hasPrint = true;
   @Input() public hasEventSelector = true;
@@ -22,6 +24,7 @@ export class CaseViewerComponent implements OnInit, OnDestroy {
   @Input() public caseDetails: CaseView;
   public caseSubscription: Subscription;
   public userAccessType: string;
+  public accessGranted: boolean;
 
   constructor(
     private readonly route: ActivatedRoute,
@@ -53,11 +56,12 @@ export class CaseViewerComponent implements OnInit, OnDestroy {
   }
 
   public setUserAccessType(caseDetails: CaseView): void {
-    // remove once CCD starts sending CAM data or Access management goes live
-    this.setMockData(caseDetails);
     if (caseDetails && Array.isArray(caseDetails.metadataFields)) {
       const access_process = caseDetails.metadataFields.find(metadataField =>
         metadataField.id === CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID);
+      const access_granted = caseDetails.metadataFields.find(metadataField =>
+        metadataField.id === CaseViewerComponent.METADATA_FIELD_ACCESS_GRANTED_ID);
+        this.accessGranted = access_granted ? access_granted.value !== CaseViewerComponent.BASIC_USER_ACCESS_TYPES : false;
       this.userAccessType = access_process ? access_process.value : null;
     }
   }
@@ -68,31 +72,8 @@ export class CaseViewerComponent implements OnInit, OnDestroy {
 
   public hasStandardAccess(): boolean {
     const featureToggleOn = this.appConfig.getAccessManagementMode();
-    return featureToggleOn ? CaseViewerComponent.NON_STANDARD_USER_ACCESS_TYPES.indexOf(this.userAccessType) === -1 : true;
-  }
-
-  // remove once Access management goes live
-  private setMockData(caseDetails: CaseView): void {
-    const accessManagementBasicViewMock = this.appConfig.getAccessManagementBasicViewMock()
-    if (accessManagementBasicViewMock && accessManagementBasicViewMock.active && !caseDetails.basicFields) {
-      const access_process_index = caseDetails.metadataFields.findIndex(metadataField =>
-        metadataField.id === CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID);
-
-      if (access_process_index > -1) {
-        caseDetails.metadataFields[access_process_index].value = accessManagementBasicViewMock.accessProcess;
-      } else {
-        const access_process: CaseField = new CaseField();
-        access_process.id = CaseViewerComponent.METADATA_FIELD_ACCESS_PROCEES_ID;
-        access_process.value = accessManagementBasicViewMock.accessProcess;
-        access_process.field_type = {
-          id: '',
-          type: 'Text'
-        };
-        caseDetails.metadataFields.push(access_process);
-      }
-
-      caseDetails.basicFields = accessManagementBasicViewMock.basicFields;
-
-    }
+    return featureToggleOn ?
+            !this.accessGranted ? CaseViewerComponent.NON_STANDARD_USER_ACCESS_TYPES.indexOf(this.userAccessType) === -1 : true
+            : true;
   }
 }