+++ src/shared/domain/definition/field-type-enum.model.ts
@@ -25,4 +25,5 @@ export type FieldTypeEnum =
   | 'CaseHistoryViewer'
   | 'Organisation'
   | 'WaysToPay'
-  | 'CaseLink';
+  | 'CaseLink'
+  | 'JudicialUserField';